Split out properties, actions and events tests from things tests
