+++ src/test/integration/things-test.ts
@@ -1,10 +1,8 @@
-import { server, httpServer, chai, mockAdapter } from '../common';
+import { server, chai, mockAdapter } from '../common';
 import { TEST_USER, createUser, headerAuth } from '../user';
 import e2p from 'event-to-promise';
 import { webSocketOpen, webSocketRead, webSocketSend, webSocketClose } from '../websocket-util';
 import WebSocket from 'ws';
-import EventSource from 'eventsource';
-import { AddressInfo } from 'net';
 import * as Constants from '../../constants';
 import Event from '../../models/event';
 import Events from '../../models/events';
@@ -24,49 +22,6 @@ const TEST_THING = {
       '@type': 'LevelProperty',
       type: 'number',
       value: 20,
-    },
-  },
-};
-
-const VALIDATION_THING = {
-  id: 'validation-1',
-  title: 'validation-1',
-  '@context': ['https://www.w3.org/2022/wot/td/v1.1', 'https://webthings.io/schemas'],
-  properties: {
-    readOnlyProp: {
-      type: 'boolean',
-      readOnly: true,
-      value: true,
-    },
-    minMaxProp: {
-      type: 'number',
-      minimum: 10,
-      maximum: 20,
-      value: 15,
-    },
-    enumProp: {
-      type: 'string',
-      enum: ['val1', 'val2', 'val3'],
-      value: 'val2',
-    },
-    multipleProp: {
-      type: 'integer',
-      minimum: 0,
-      maximum: 600,
-      value: 10,
-      multipleOf: 5,
-    },
-  },
-};
-
-const EVENT_THING = {
-  id: 'event-thing1',
-  title: 'Event Thing',
-  '@context': ['https://www.w3.org/2022/wot/td/v1.1', 'https://webthings.io/schemas'],
-  events: {
-    overheated: {
-      type: 'number',
-      unit: 'degree celsius',
     },
   },
 };
@@ -340,165 +295,6 @@ describe('things/', function () {
     expect(res.status).toEqual(200);
     expect(res.body).toHaveProperty('title');
     expect(res.body.title).toEqual('new title');
-  });
-
-  it('GET all properties of a thing', async () => {
-    await addDevice();
-    const res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/test-1/properties`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toHaveProperty('power');
-    expect(res.body.power).toEqual(false);
-    expect(res.body).toHaveProperty('percent');
-    expect(res.body.percent).toEqual(20);
-  });
-
-  it('GET a property of a thing', async () => {
-    await addDevice();
-    const res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/test-1/properties/power`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual(false);
-  });
-
-  it('fail to GET a nonexistent property of a thing', async () => {
-    await addDevice();
-    const err = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/test-1/properties/xyz`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(err.status).toEqual(500);
-  });
-
-  it('fail to GET a property of a nonexistent thing', async () => {
-    const err = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/test-1a/properties/power`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-    expect(err.status).toEqual(500);
-  });
-
-  it('fail to set a property of a thing', async () => {
-    await addDevice();
-    const err = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/test-1/properties/power`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send();
-    expect(err.status).toEqual(400);
-  });
-
-  it('fail to set a property of a thing', async () => {
-    const err = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/test-1/properties/power`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send('foo');
-    expect(err.status).toEqual(400);
-  });
-
-  it('set a property of a thing', async () => {
-    // Set it to true
-    await addDevice();
-    const on = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/test-1/properties/power`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(JSON.stringify(true));
-
-    expect(on.status).toEqual(204);
-
-    // Check that it was set to true
-    const readOn = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/test-1/properties/power`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(readOn.status).toEqual(200);
-    expect(readOn.body).toEqual(true);
-
-    // Set it back to false
-    const off = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/test-1/properties/power`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(JSON.stringify(false));
-
-    expect(off.status).toEqual(204);
-
-    // Check that it was set to false
-    const readOff = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/test-1/properties/power`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(readOff.status).toEqual(200);
-    expect(readOff.body).toEqual(false);
-  });
-
-  it('set multiple properties of a thing', async () => {
-    // Set properties
-    await addDevice();
-    const setProperties = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/test-1/properties`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(
-        JSON.stringify({
-          power: true,
-          percent: 42,
-        })
-      );
-
-    expect(setProperties.status).toEqual(204);
-
-    // Check that the properties were set
-    const getProperties = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/test-1/properties`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(getProperties.status).toEqual(200);
-    expect(getProperties.body.power).toEqual(true);
-    expect(getProperties.body.percent).toEqual(42);
-  });
-
-  it('fail to set multiple properties of a thing', async () => {
-    // Set properties
-    await addDevice();
-    const setProperties = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/test-1/properties`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(
-        JSON.stringify({
-          power: true,
-          percent: 42,
-          invalidpropertyname: true,
-        })
-      );
-
-    expect(setProperties.status).toEqual(500);
   });
 
   it('fail to set x and y coordinates of a non-existent thing', async () => {
@@ -1026,157 +822,6 @@ describe('things/', function () {
     await webSocketClose(ws);
   });
 
-  it('should be able to retrieve events', async () => {
-    await addDevice();
-
-    let res = await chai
-      .request(server)
-      .get(Constants.EVENTS_PATH)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(Array.isArray(res.body)).toBeTruthy();
-    expect(res.body.length).toEqual(0);
-
-    res = await chai
-      .request(server)
-      .get(`${Constants.EVENTS_PATH}/a`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(Array.isArray(res.body)).toBeTruthy();
-    expect(res.body.length).toEqual(0);
-
-    const thingBase = `${Constants.THINGS_PATH}/${TEST_THING.id}`;
-
-    res = await chai
-      .request(server)
-      .get(thingBase + Constants.EVENTS_PATH)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(Array.isArray(res.body)).toBeTruthy();
-    expect(res.body.length).toEqual(0);
-
-    res = await chai
-      .request(server)
-      .get(`${thingBase}${Constants.EVENTS_PATH}/a`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(Array.isArray(res.body)).toBeTruthy();
-    expect(res.body.length).toEqual(0);
-
-    const eventA = new Event('a', 'just a cool event', TEST_THING.id);
-    const eventB = new Event('b', 'just a boring event', TEST_THING.id);
-    await Events.add(eventA);
-    await Events.add(eventB);
-
-    res = await chai
-      .request(server)
-      .get(thingBase + Constants.EVENTS_PATH)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(Array.isArray(res.body)).toBeTruthy();
-    expect(res.body.length).toEqual(2);
-    expect(res.body[0]).toHaveProperty('a');
-    expect(res.body[0].a).toHaveProperty('data');
-    expect(res.body[0].a.data).toBe('just a cool event');
-    expect(res.body[0].a).toHaveProperty('timestamp');
-    expect(res.body[1]).toHaveProperty('b');
-    expect(res.body[1].b).toHaveProperty('data');
-    expect(res.body[1].b.data).toBe('just a boring event');
-    expect(res.body[1].b).toHaveProperty('timestamp');
-
-    res = await chai
-      .request(server)
-      .get(`${thingBase}${Constants.EVENTS_PATH}/a`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(Array.isArray(res.body)).toBeTruthy();
-    expect(res.body.length).toEqual(1);
-    expect(res.body[0]).toHaveProperty('a');
-    expect(res.body[0].a).toHaveProperty('data');
-    expect(res.body[0].a.data).toBe('just a cool event');
-    expect(res.body[0].a).toHaveProperty('timestamp');
-  });
-
-  it('should be able to subscribe to an event using EventSource', async () => {
-    await addDevice(EVENT_THING);
-    const address = <AddressInfo>httpServer.address();
-
-    const eventSourceURL =
-      `http://127.0.0.1:${address.port}${Constants.THINGS_PATH}/` +
-      `${EVENT_THING.id}/events/overheated?jwt=${jwt}`;
-    const eventSource = new EventSource(eventSourceURL) as EventTarget & EventSource;
-    await e2p(eventSource, 'open');
-    const overheatedEvent = new Event('overheated', 101, EVENT_THING.id);
-    const [, event] = await Promise.all([
-      Events.add(overheatedEvent),
-      e2p(eventSource, 'overheated'),
-    ]);
-    expect(event.type).toEqual('overheated');
-    expect(JSON.parse(event.data)).toEqual(101);
-    eventSource.close();
-  });
-
-  it('should be able to subscribe to all events on a thing using EventSource', async () => {
-    await addDevice(EVENT_THING);
-    const address = <AddressInfo>httpServer.address();
-
-    const eventSourceURL =
-      `http://127.0.0.1:${address.port}${Constants.THINGS_PATH}/` +
-      `${EVENT_THING.id}/events?jwt=${jwt}`;
-    const eventsSource = new EventSource(eventSourceURL) as EventTarget & EventSource;
-    await e2p(eventsSource, 'open');
-    const overheatedEvent2 = new Event('overheated', 101, EVENT_THING.id);
-    const [, event2] = await Promise.all([
-      Events.add(overheatedEvent2),
-      e2p(eventsSource, 'overheated'),
-    ]);
-    expect(event2.type).toEqual('overheated');
-    expect(JSON.parse(event2.data)).toEqual(101);
-    eventsSource.close();
-  });
-
-  it('should not be able to subscribe events on a thing that doesnt exist', async () => {
-    await addDevice(EVENT_THING);
-    const address = <AddressInfo>httpServer.address();
-
-    const eventSourceURL =
-      `http://127.0.0.1:${address.port}${Constants.THINGS_PATH}` +
-      `/non-existent-thing/events/overheated?jwt=${jwt}`;
-    const thinglessEventSource = new EventSource(eventSourceURL) as EventTarget & EventSource;
-    thinglessEventSource.onerror = jest.fn();
-    thinglessEventSource.onopen = jest.fn();
-    await e2p(thinglessEventSource, 'error');
-    expect(thinglessEventSource.onopen).not.toBeCalled();
-    expect(thinglessEventSource.onerror).toBeCalled();
-  });
-
-  it('should not be able to subscribe to an event that doesnt exist', async () => {
-    await addDevice(EVENT_THING);
-    const address = <AddressInfo>httpServer.address();
-
-    const eventSourceURL =
-      `http://127.0.0.1:${address.port}${Constants.THINGS_PATH}` +
-      `${EVENT_THING.id}/events/non-existentevent?jwt=${jwt}`;
-    const eventlessEventSource = new EventSource(eventSourceURL) as EventTarget & EventSource;
-    eventlessEventSource.onerror = jest.fn();
-    eventlessEventSource.onopen = jest.fn();
-    await e2p(eventlessEventSource, 'error');
-    expect(eventlessEventSource.onopen).not.toBeCalled();
-    expect(eventlessEventSource.onerror).toBeCalled();
-  });
-
   // eslint-disable-next-line @typescript-eslint/quotes
   it("should receive thing's action status messages over websocket", async () => {
     await addDevice();
@@ -1301,47 +946,6 @@ describe('things/', function () {
       .set(...headerAuth(jwt));
     expect(res.status).toEqual(200);
     expect(Object.keys(res.body).length).toEqual(0);
-  });
-
-  it('fails to create an action on a nonexistent thing', async () => {
-    const thingBase = `${Constants.THINGS_PATH}/nonexistent-thing`;
-
-    const input = {
-      timeout: 60,
-    };
-
-    const err = await chai
-      .request(server)
-      .post(`${thingBase}${Constants.ACTIONS_PATH}/pair`)
-      .set(...headerAuth(jwt))
-      .set('Accept', 'application/json')
-      .send(input);
-    expect(err.status).toEqual(404);
-  });
-
-  it('fails to create thing action which does not exist', async () => {
-    await addDevice(piDescr);
-
-    const thingBase = `${Constants.THINGS_PATH}/${piDescr.id}`;
-
-    const res = await chai
-      .request(server)
-      .get(thingBase)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-    expect(res.status).toEqual(200);
-
-    const input = {
-      timeout: 60,
-    };
-
-    const err = await chai
-      .request(server)
-      .post(`${thingBase}${Constants.ACTIONS_PATH}/pair`)
-      .set(...headerAuth(jwt))
-      .set('Accept', 'application/json')
-      .send(input);
-    expect(err.status).toEqual(400);
   });
 
   it('should create an action over websocket', async () => {
@@ -1489,155 +1093,5 @@ describe('things/', function () {
     expect(res.status).toEqual(200);
     expect(res.body).toHaveProperty('title');
     expect(res.body.title).toEqual(piDescr.title);
-  });
-
-  it('fail to set read-only property', async () => {
-    await addDevice(VALIDATION_THING);
-
-    let res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/readOnlyProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual(true);
-
-    const err = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/validation-1/properties/readOnlyProp`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(JSON.stringify(false));
-    expect(err.status).toEqual(400);
-
-    res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/readOnlyProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual(true);
-  });
-
-  it('fail to set invalid number property value', async () => {
-    await addDevice(VALIDATION_THING);
-
-    let res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/minMaxProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual(15);
-
-    let err = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/validation-1/properties/minMaxProp`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(JSON.stringify(0));
-    expect(err.status).toEqual(400);
-
-    res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/minMaxProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual(15);
-
-    err = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/validation-1/properties/minMaxProp`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(JSON.stringify(30));
-    expect(err.status).toEqual(400);
-
-    res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/minMaxProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual(15);
-
-    res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/multipleProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual(10);
-
-    err = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/validation-1/properties/multipleProp`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(JSON.stringify(3));
-    expect(err.status).toEqual(400);
-
-    res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/multipleProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual(10);
-
-    res = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/validation-1/properties/multipleProp`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(JSON.stringify(30));
-    expect(res.status).toEqual(204);
-
-    res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/multipleProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual(30);
-  });
-
-  it('fail to set invalid enum property value', async () => {
-    await addDevice(VALIDATION_THING);
-
-    let res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/enumProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual('val2');
-
-    const err = await chai
-      .request(server)
-      .put(`${Constants.THINGS_PATH}/validation-1/properties/enumProp`)
-      .type('json')
-      .set(...headerAuth(jwt))
-      .send(JSON.stringify('val0'));
-    expect(err.status).toEqual(400);
-
-    res = await chai
-      .request(server)
-      .get(`${Constants.THINGS_PATH}/validation-1/properties/enumProp`)
-      .set('Accept', 'application/json')
-      .set(...headerAuth(jwt));
-
-    expect(res.status).toEqual(200);
-    expect(res.body).toEqual('val2');
   });
 });
